+++ src/cljs/rems/administration/catalogue_items.cljs
@@ -1,5 +1,6 @@
 (ns rems.administration.catalogue-items
-  (:require [re-frame.core :as rf]
+  (:require [cljs-time.coerce :as time-coerce]
+            [re-frame.core :as rf]
             [rems.administration.administration :refer [administration-navigator-container]]
             [rems.administration.catalogue-item :as catalogue-item]
             [rems.administration.status-flags :as status-flags]
@@ -9,12 +10,13 @@
             [rems.spinner :as spinner]
             [rems.table :as table]
             [rems.text :refer [localize-time text get-localized-title]]
-            [rems.util :refer [fetch put!]]))
+            [rems.util :refer [navigate! fetch put!]] ))
 
 (rf/reg-event-fx
  ::enter-page
  (fn [{:keys [db]}]
-   {:dispatch-n [[::fetch-catalogue]
+   {:db (assoc db ::selected-items (or (::selected-items db) #{}))
+    :dispatch-n [[::fetch-catalogue]
                  [:rems.table/reset]]}))
 
 (rf/reg-event-fx
@@ -60,12 +62,30 @@
           :error-handler (flash-message/default-error-handler :top description)})
    {}))
 
-(defn- to-create-catalogue-item []
+(rf/reg-event-db
+ ::set-selected-items
+ (fn [db [_ items]]
+   (assoc db ::selected-items items)))
+
+(rf/reg-sub
+ ::selected-items
+ (fn [db _]
+   (::selected-items db)))
+
+(defn- create-catalogue-item-button []
   [atoms/link {:class "btn btn-primary"}
    "/administration/catalogue-items/create"
    (text :t.administration/create-catalogue-item)])
 
-(defn- to-catalogue-item [catalogue-item-id]
+(defn- change-form-button [items]
+  [:button.btn.btn-primary
+   {:disabled (when (empty? items) :disabled)
+    :on-click (fn []
+                (rf/dispatch [:rems.administration.change-catalogue-item-form/enter-page items])
+                (navigate! "/administration/catalogue-items/change-form") )}
+   (text :t.administration/change-form)])
+
+(defn- view-button [catalogue-item-id]
   [atoms/link {:class "btn btn-primary"}
    (str "/administration/catalogue-items/" catalogue-item-id)
    (text :t.administration/view)])
@@ -78,7 +98,9 @@
  (fn [[catalogue language] _]
    (map (fn [item]
           {:key (:id item)
-           :name {:value (get-localized-title item language)}
+           :name {:value (get-localized-title item language)
+                  :sort-value [(get-localized-title item language)
+                               (- (time-coerce/to-long (:start item)))]} ; secondary sort by created, reverse
            :resource (let [value (:resource-name item)]
                        {:value value
                         :td [:td.resource
@@ -105,7 +127,7 @@
                            [readonly-checkbox {:value checked?}]]
                       :sort-value (if checked? 1 2)})
            :commands {:td [:td.commands
-                           [to-catalogue-item (:id item)]
+                           [view-button (:id item)]
                            [roles/when roles/show-admin-edit-buttons?
                             [catalogue-item/edit-button (:id item)]
                             [status-flags/enabled-toggle item #(rf/dispatch [::set-catalogue-item-enabled %1 %2 [::fetch-catalogue]])]
@@ -131,10 +153,15 @@
                                     :sortable? false
                                     :filterable? false}]
                          :rows [::catalogue-table-rows]
-                         :default-sort-column :name}]
+                         :default-sort-column :name
+                         :selectable? true
+                         :on-select #(rf/dispatch [::set-selected-items %])}]
     [:div.mt-3
      [table/search catalogue-table]
      [table/table catalogue-table]]))
+
+(defn- items-by-ids [items ids]
+  (filter (comp ids :id) items))
 
 (defn catalogue-items-page []
   (into [:div
@@ -144,7 +171,10 @@
         (if @(rf/subscribe [::loading?])
           [[spinner/big]]
           [[roles/when roles/show-admin-edit-buttons?
-            [to-create-catalogue-item]
-            [status-flags/display-archived-toggle #(rf/dispatch [::fetch-catalogue])]
+            [:div.commands.text-left.pl-0
+             [create-catalogue-item-button]
+             [change-form-button (items-by-ids @(rf/subscribe [::catalogue]) @(rf/subscribe [::selected-items]))]]
+            [status-flags/display-archived-toggle #(do (rf/dispatch [::fetch-catalogue])
+                                                       (rf/dispatch [:rems.table/set-selected-rows {:id ::catalogue} nil]))]
             [status-flags/disabled-and-archived-explanation]]
            [catalogue-list]])))