+++ test/clj/rems/api/test_catalogue_items.clj
@@ -128,7 +128,7 @@
       (is (response-is-unauthorized? response))
       (is (str/includes? body "Invalid anti-forgery token"))))
   (testing "create with wrong API-Key"
-    (is (= "invalid api key"
+    (is (= "unauthorized"
            (-> (request :post (str "/api/catalogue-items/create"))
                (assoc-in [:headers "x-rems-api-key"] "invalid-api-key")
                (json-body {:form 1
@@ -146,8 +146,8 @@
       (is (response-is-unauthorized? response))
       (is (str/includes? body "Invalid anti-forgery token"))))
   (testing "edit with wrong API-Key"
-    (is (= "invalid api key"
-           (-> (request :post (str "/api/catalogue-items/edit"))
+    (is (= "unauthorized"
+           (-> (request :put (str "/api/catalogue-items/edit"))
                (assoc-in [:headers "x-rems-api-key"] "invalid-api-key")
                (json-body {:id 1
                            :localizations {:en {:title "malicious localization"}}})
@@ -173,7 +173,7 @@
                                       (json-body {:form new-form-id})
                                       handler
                                       read-ok-body
-                                      :new-catalogue-item-id)
+                                      :catalogue-item-id)
             new-catalogue-item (-> (request :get (str "/api/catalogue-items/" new-catalogue-item-id))
                                    (authenticate api-key "owner")
                                    handler